+++ packages/angular/src/lib/blok/sb-blok.feature.ts
@@ -1,5 +1,5 @@
 import { Injectable, inject } from '@angular/core';
-import { STORYBLOK_COMPONENTS, BaseComponentResolver } from './components.feature';
+import { STORYBLOK_COMPONENTS, BaseComponentResolver } from '../components.feature';
 
 /** Resolver for Storyblok blok components. */
 @Injectable({ providedIn: 'root' })
