+++ packages/angular/src/lib/types.ts
@@ -8,7 +8,7 @@ export interface SbBlokData extends ISbComponentType<string> {
 }
 
 //TODO: This is a temporary this should come from the rich text package
-export type RichTextBaseProps<T extends TiptapComponentName> = T extends PMNode['type']
+type RichTextBaseProps<T extends TiptapComponentName> = T extends PMNode['type']
   ? Extract<
       PMNode,
       {
@@ -23,3 +23,5 @@ export type RichTextBaseProps<T extends TiptapComponentName> = T extends PMNode[
         }
       >
     : never;
+
+export type RichTextComponentProps<T extends TiptapComponentName> = RichTextBaseProps<T>;
